+++ src/Sentry.Unity.Editor/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("Sentry.Unity.Editor.iOS")]
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor.Tests")]