+++ src/Sentry.Unity.Editor/SentryWindow.cs
@@ -22,11 +22,12 @@ namespace Sentry.Unity.Editor
         protected virtual string SentryOptionsAssetName { get; } = ScriptableSentryUnityOptions.ConfigName;
 
         public ScriptableSentryUnityOptions Options { get; private set; } = null!; // Set by OnEnable()
+        public SentryCliOptions CliOptions { get; private set; } = null!; // Set by OnEnable()
 
         public event Action<ValidationError> OnValidationError = _ => { };
 
         private int _currentTab = 0;
-        private string[] _tabs = new[] { "Core", "Enrichment", "Transport", "Advanced" };
+        private string[] _tabs = new[] { "Core", "Enrichment", "Transport", "Advanced", "Editor" };
 
         private void OnEnable()
         {
@@ -35,6 +36,7 @@ namespace Sentry.Unity.Editor
 
             CheckForAndConvertJsonConfig();
             Options = LoadOptions();
+            CliOptions = SentryCliOptions.LoadCliOptions();
         }
 
         private ScriptableSentryUnityOptions LoadOptions()
@@ -131,6 +133,9 @@ namespace Sentry.Unity.Editor
                 case 3:
                     DisplayDebug();
                     break;
+                case 4:
+                    DisplayEditor();
+                    break;
                 default:
                     break;
             }
@@ -150,6 +155,11 @@ namespace Sentry.Unity.Editor
             Options.CaptureInEditor = EditorGUILayout.Toggle(
                 new GUIContent("Capture In Editor", "Capture errors while running in the Editor."),
                 Options.CaptureInEditor);
+
+            Options.EnableLogDebouncing = EditorGUILayout.Toggle(
+                new GUIContent("Enable Log Debouncing", "The SDK debounces log messages of the same type if " +
+                                                    "they are more frequent than once per second."),
+                Options.EnableLogDebouncing);
 
             EditorGUILayout.Space();
             EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
@@ -341,6 +351,33 @@ namespace Sentry.Unity.Editor
                 new GUIContent("Android Native Support", "Whether to enable Native Android support to " +
                                                          "capture errors written in languages such as Java, Kotlin, C and C++."),
                 Options.AndroidNativeSupportEnabled);
+        }
+
+        private void DisplayEditor()
+        {
+            CliOptions.UploadSymbols = EditorGUILayout.BeginToggleGroup(
+                new GUIContent("Upload Symbols", "Whether debug symbols should be uploaded automatically " +
+                                                 "on release builds."),
+                CliOptions.UploadSymbols);
+
+            CliOptions.UploadDevelopmentSymbols = EditorGUILayout.Toggle(
+                new GUIContent("Upload Dev Symbols", "Whether debug symbols should be uploaded automatically " +
+                                                     "on development builds."),
+                CliOptions.UploadDevelopmentSymbols);
+
+            EditorGUILayout.EndToggleGroup();
+
+            CliOptions.Auth = EditorGUILayout.TextField(
+                new GUIContent("Auth Token", "The authorization token from your user settings in Sentry"),
+                CliOptions.Auth);
+
+            CliOptions.Organization = EditorGUILayout.TextField(
+                new GUIContent("Org Slug", "The organization slug in Sentry"),
+                CliOptions.Organization);
+
+            CliOptions.Project = EditorGUILayout.TextField(
+                new GUIContent("Project Name", "The project name in Sentry"),
+                CliOptions.Project);
         }
 
         private void OnLostFocus()