+++ CHANGELOG.md
@@ -1,13 +1,38 @@
 # Changelog
 
-## Unreleased
+## 0.7.0
+
+### Features
+
+- Added automated symbols upload support for Android ([#408](https://github.com/getsentry/sentry-unity/pull/408))
+- Avoid DLL conflict with other Unity packages ([#425](https://github.com/getsentry/sentry-unity/issues/425))
+
+## 0.6.2
+
+- fix release packaging ([#417](https://github.com/getsentry/sentry-unity/pull/417))
+
+## 0.6.1
+
+### Features
+
+- Added sentry-cli to Unity package ([#414](https://github.com/getsentry/sentry-unity/pull/414))
 
 ### Fixes
 
+- Added missing release string validation ([#389](https://github.com/getsentry/sentry-unity/pull/389))
+- Sentry internal logs no longer show up as breadcrumbs ([#377](https://github.com/getsentry/sentry-unity/pull/377))
 - Fixed missing context data when initializing SDK programmatically ([#376](https://github.com/getsentry/sentry-unity/pull/376))
 - Fixed CaptureInEditor flag when initializing SDK programmatically ([#370](https://github.com/getsentry/sentry-unity/pull/370))
 - Preventing numeric options to be set negative in the editor window ([#364](https://github.com/getsentry/sentry-unity/pull/364))
-
+- Bump Sentry .NET SDK 3.11.1 ([#407](https://github.com/getsentry/sentry-unity/pull/407))
+  - [changelog](https://github.com/getsentry/sentry-dotnet/blob/3.11.1/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-dotnet/compare/3.9.3...3.11.1)
+- Bump Sentry Cocoa SDK 7.5.2 ([#407](https://github.com/getsentry/sentry-unity/pull/407))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/7.5.2/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/7.3.0...7.5.2)
+- Bump Sentry Android SDK 5.4.0 ([#411](https://github.com/getsentry/sentry-unity/pull/411))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/5.4.0/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-java/compare/5.2.1...5.4.0)
 
 ## 0.6.0